fix: split client and server code
