+++ integrations/disco/hooks/use-disco-get-profile-from-address.ts
@@ -1,7 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
-import error from 'next/error'
 
-import { appDiscoGetProfileFromAddress } from '@/integrations/disco/routes/get-profile-from-address'
+import { appDiscoGetProfileFromAddress } from '@/integrations/disco/routes/get-profile-from-address/client'
 
 export const useDiscoGetProfileFromAddress = (address?: `0x${string}`, queryKey?: any) => {
   return useQuery(['discoProfileFromAddress', address, queryKey], async () => appDiscoGetProfileFromAddress(address))
