+++ integrations/disco/routes/get-profile-from-address/client.ts
@@ -1,0 +1,14 @@
+import axios from 'axios'
+
+export async function appDiscoGetProfileFromAddress(address?: `0x${string}`): Promise<any> {
+  try {
+    const { data } = await axios.get(`/api/disco/profile-from-address`, {
+      params: {
+        address: address,
+      },
+    })
+    return data
+  } catch (error: any) {
+    throw error
+  }
+}
