+++ integrations/disco/routes/get-profile-from-address/index.ts
@@ -1,5 +1,3 @@
-import axios from 'axios'
-
 import { discoClient } from '@/integrations/disco/disco-client'
 
 export async function discoGetProfileFromAddress(address?: string): Promise<any> {
@@ -8,19 +6,6 @@ export async function discoGetProfileFromAddress(address?: string): Promise<any>
       return null
     }
     const { data } = await discoClient.get(`/profile/address/${address}`)
-    return data
-  } catch (error: any) {
-    throw error
-  }
-}
-
-export async function appDiscoGetProfileFromAddress(address?: `0x${string}`): Promise<any> {
-  try {
-    const { data } = await axios.get(`/api/disco/profile-from-address`, {
-      params: {
-        address: address,
-      },
-    })
     return data
   } catch (error: any) {
     throw error
