+++ integrations/disco/routes/get-profile-from-did/index.ts
@@ -1,0 +1,6 @@
+import { discoClient } from '@/integrations/disco/disco-client'
+
+export async function discoGetProfileFromDID(did: string): Promise<any> {
+  const { data } = await discoClient.get(`/profile/${did}`)
+  return data
+}
