+++ integrations/etherscan/hooks/use-etherscan-account-transactions.ts
@@ -1,6 +1,6 @@
 import { useQuery } from 'wagmi'
 
-import { appEtherscanAccountTransactions } from '@/integrations/etherscan/actions/etherscan-account-transactions'
+import { appEtherscanAccountTransactions } from '@/integrations/etherscan/actions/etherscan-account-transactions/client'
 
 export const useEtherscanAccountTransactions = (params?: BlockPagination, queryKey?: any) => {
   return useQuery(['accountTransactions', params, queryKey], () => appEtherscanAccountTransactions(params), {
